refactor: use encoding/binary for position in txIndex

+++ wallet/wallet_test.go
@@ -27,7 +27,7 @@ import (
 func TestEncodeDecodeGlobalTxIndex(t *testing.T) {
 	want := &struct {
 		BlockHash bc.Hash
-		Position  int
+		Position  uint64
 	}{
 		BlockHash: bc.NewHash([32]byte{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f, 0x20}),
 		Position:  1,
@@ -199,7 +199,7 @@ func TestWalletUpdate(t *testing.T) {
 	for position, tx := range block.Transactions {
 		get := w.DB.Get(calcGlobalTxIndexKey(tx.ID.String()))
 		bh := block.BlockHeader.Hash()
-		expect := calcGlobalTxIndex(&bh, position)
+		expect := calcGlobalTxIndex(&bh, uint64(position))
 		if !reflect.DeepEqual(get, expect) {
 			t.Fatalf("position#%d: compare retrieved globalTxIdx err", position)
 		}
